+++ CHANGES.rst
@@ -10,3 +10,5 @@ CHANGES
 
 - Dropped: Adding `sub app` via `app.router.add_subapp()` is deprecated
   use `app.add_subapp()` instead #1592
+
+- Dropped: `Application.finish()` and `Application.register_on_finish()`  #1602